simd_json::borrowed

Enum Value

pub enum Value<'value> {
    Static(StaticNode),
    String(Cow<'value, str>),
    Array(Box<Vec<Value<'value>>>),
    Object(Box<Object<'value>>),
}

Borrowed JSON-DOM Value, consider using the ValueTrait to access its content

Variants

Static(StaticNode)

Static values

String(Cow<'value, str>)

string type

Array(Box<Vec<Value<'value>>>)

array type

Object(Box<Object<'value>>)

object type

Implementations

impl<'value> Value<'value>

pub fn into_static(self) -> Value<'static>

Enforces static lifetime on a borrowed value, this will force all strings to become owned COW’s, the same applies for Object keys.

pub fn clone_static(&self) -> Value<'static>

Clones the current value and enforces a static lifetime, it works the same as into_static but includes cloning logic