simd_json::lazy
Enum Value

pub enum Value<'borrow, 'tape, 'input> {
    Tape(Value<'tape, 'input>),
    Value(Cow<'borrow, Value<'input>>),
}

A lazy value, this gets initialized with a tape and as long as only non mutating operations are performed it will stay a tape. If a mutating operation is performed it will upgrade to a borrowed value.

Variants

Tape(Value<'tape, 'input>)

tape variant

Value(Cow<'borrow, Value<'input>>)

borrowed variant

Implementations

impl<'borrow, 'tape, 'input> Value<'borrow, 'tape, 'input>
where 'input: 'tape,

pub fn is_array(&self) -> bool

returns true if the current value can be represented as an array

pub fn is_object(&self) -> bool

returns true if the current value can be represented as an object

impl<'borrow, 'tape, 'input> Value<'borrow, 'tape, 'input>

pub fn try_get<Q>( &self, k: &Q, ) -> Result<Option<Value<'_, 'tape, 'input>>, TryTypeError>
where str: Borrow<Q> + Hash + Eq, for<'b> Cow<'b, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get a value based on a key, returns a TryTypeError if the current Value isn’t an Object, returns None if the key isn’t in the object

Errors

if the value is not an object

impl<'borrow, 'tape, 'input> Value<'borrow, 'tape, 'input>
where 'input: 'tape,

pub fn try_get_idx( &self, i: usize, ) -> Result<Option<Value<'_, 'tape, 'input>>, TryTypeError>

Tries to get a value based on n index, returns a type error if the current value isn’t an Array, returns None if the index is out of bounds

Errors

if the requested type doesn’t match the actual type or the value is not an object

impl<'borrow, 'tape, 'input> Value<'borrow, 'tape, 'input>
where 'input: 'tape,

pub fn as_array(&self) -> Option<Array<'_, 'tape, 'input>>

Tries to represent the value as an array and returns a reference to it

pub fn as_object(&self) -> Option<Object<'_, 'tape, 'input>>

Tries to represent the value as an array and returns a reference to it

impl<'borrow, 'tape, 'input> Value<'borrow, 'tape, 'input>

pub fn try_as_array(&self) -> Result<Array<'_, 'tape, 'input>, TryTypeError>

Tries to represent the value as an array and returns a reference to it

Errors

if the requested type doesn’t match the actual type

pub fn try_as_object(&self) -> Result<Object<'_, 'tape, 'input>, TryTypeError>

Tries to represent the value as an object and returns a reference to it

Errors

if the requested type doesn’t match the actual type

impl<'borrow, 'tape, 'input> Value<'borrow, 'tape, 'input>

pub fn get<'k, Q>(&self, k: &'k Q) -> Option<Value<'_, 'tape, 'input>>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Gets a ref to a value based on a key, returns None if the current Value isn’t an Object or doesn’t contain the key it was asked for.

pub fn contains_key<Q>(&self, k: &Q) -> bool
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Checks if a Value contains a given key. This will return flase if Value isn’t an object

impl<'borrow, 'tape, 'input> Value<'borrow, 'tape, 'input>

pub fn get_idx(&self, i: usize) -> Option<Value<'_, 'tape, 'input>>

Gets a ref to a value based on n index, returns None if the current Value isn’t an Array or doesn’t contain the index it was asked for.

impl<'borrow, 'tape, 'input> Value<'borrow, 'tape, 'input>
where 'input: 'tape,

pub fn get_bool<Q>(&self, k: &Q) -> Option<bool>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a bool

pub fn get_i128<Q>(&self, k: &Q) -> Option<i128>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a i128

pub fn get_i64<Q>(&self, k: &Q) -> Option<i64>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a i64

pub fn get_i32<Q>(&self, k: &Q) -> Option<i32>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a i32

pub fn get_i16<Q>(&self, k: &Q) -> Option<i16>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a i16

pub fn get_i8<Q>(&self, k: &Q) -> Option<i8>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a i8

pub fn get_u128<Q>(&self, k: &Q) -> Option<u128>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a u128

pub fn get_u64<Q>(&self, k: &Q) -> Option<u64>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a u64

pub fn get_usize<Q>(&self, k: &Q) -> Option<usize>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a usize

pub fn get_u32<Q>(&self, k: &Q) -> Option<u32>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a u32

pub fn get_u16<Q>(&self, k: &Q) -> Option<u16>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a u16

pub fn get_u8<Q>(&self, k: &Q) -> Option<u8>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a u8

pub fn get_f64<Q>(&self, k: &Q) -> Option<f64>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a f64

pub fn get_f32<Q>(&self, k: &Q) -> Option<f32>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a f32

pub fn get_str<Q>(&self, k: &Q) -> Option<&str>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a str

impl<'borrow, 'tape, 'input> Value<'borrow, 'tape, 'input>

pub fn get_array<Q>(&self, k: &Q) -> Option<Array<'_, 'tape, 'input>>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a array

pub fn get_object<Q>(&self, k: &Q) -> Option<Object<'_, 'tape, 'input>>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a object

impl<'borrow, 'tape, 'input> Value<'borrow, 'tape, 'input>

pub fn try_get_array<Q>( &self, k: &Q, ) -> Result<Option<Array<'_, 'tape, 'input>>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as an array, returns an error if it isn’t a array

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_object<Q>( &self, k: &Q, ) -> Result<Option<Object<'_, 'tape, 'input>>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as an object, returns an error if it isn’t an object

Errors

if the requested type doesn’t match the actual type or the value is not an object

impl<'borrow, 'tape, 'input> Value<'borrow, 'tape, 'input>

pub fn try_get_bool<Q>(&self, k: &Q) -> Result<Option<bool>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a bool, returns an error if it isn’t bool

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_i128<Q>(&self, k: &Q) -> Result<Option<i128>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a i128, returns an error if it isn’t i128

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_i64<Q>(&self, k: &Q) -> Result<Option<i64>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a i64, returns an error if it isn’t a i64

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_i32<Q>(&self, k: &Q) -> Result<Option<i32>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a i32, returns an error if it isn’t a i32

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_i16<Q>(&self, k: &Q) -> Result<Option<i16>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a i16, returns an error if it isn’t a i16

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_i8<Q>(&self, k: &Q) -> Result<Option<i8>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a u128, returns an error if it isn’t a u128

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_u128<Q>(&self, k: &Q) -> Result<Option<u128>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a u64, returns an error if it isn’t a u64

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_u64<Q>(&self, k: &Q) -> Result<Option<u64>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a usize, returns an error if it isn’t a usize

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_usize<Q>(&self, k: &Q) -> Result<Option<usize>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a u32, returns an error if it isn’t a u32

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_u32<Q>(&self, k: &Q) -> Result<Option<u32>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a u16, returns an error if it isn’t a u16

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_u16<Q>(&self, k: &Q) -> Result<Option<u16>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a u8, returns an error if it isn’t a u8

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_u8<Q>(&self, k: &Q) -> Result<Option<u8>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a u8, returns an error if it isn’t a u8

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_f64<Q>(&self, k: &Q) -> Result<Option<f64>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a f64, returns an error if it isn’t a f64

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_f32<Q>(&self, k: &Q) -> Result<Option<f32>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a f32, returns an error if it isn’t a f32

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_str<Q>(&self, k: &Q) -> Result<Option<&str>, TryTypeError>
where str: Borrow<Q>, for<'a> Cow<'a, str>: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a str, returns an error if it isn’t a str

Errors

if the requested type doesn’t match the actual type or the value is not an object

impl<'borrow, 'tape, 'input> Value<'borrow, 'tape, 'input>

pub fn into_value(self) -> Value<'input>

turns the lazy value into a borrowed value

pub fn into_owned<'snot>(self) -> Value<'snot, 'tape, 'input>

extends the Value COW is owned

pub fn is_tape(&self) -> bool

returns true when the current representation is a tape

pub fn is_value(&self) -> bool

returns true when the current representation is a borrowed value this is the opposite of is_tape

pub fn from_tape(tape: Value<'tape, 'input>) -> Self

Creates a new lazy Value from a tape