simd_json::owned

Function to_value

pub fn to_value(s: &mut [u8]) -> Result<Value>

Parses a slice of bytes into a Value dom.

This function will rewrite the slice to de-escape strings. We do not keep any references to the raw data but re-allocate owned memory wherever required thus returning a value without a lifetime.

Errors

Will return Err if s is invalid JSON.