Trait simd_json::prelude::derived::ValueObjectAccessAsContainer

pub trait ValueObjectAccessAsContainer {
    type Key: ?Sized;
    type Target;
    type Array: Array;
    type Object: Object;

    // Required methods
    fn get_array<Q>(&self, k: &Q) -> Option<&Self::Array>
       where Self::Key: Borrow<Q> + Hash + Eq,
             Q: Hash + Eq + Ord + ?Sized;
    fn get_object<Q>(&self, k: &Q) -> Option<&Self::Object>
       where Self::Key: Borrow<Q> + Hash + Eq,
             Q: Hash + Eq + Ord + ?Sized;
}

Access to container values in an object

Required Associated Types

type Key: ?Sized

The type for Objects

type Target

The target for nested lookups

type Array: Array

The array structure

type Object: Object

The object structure

Required Methods

fn get_array<Q>(&self, k: &Q) -> Option<&Self::Array>
where Self::Key: Borrow<Q> + Hash + Eq, Q: Hash + Eq + Ord + ?Sized,

Tries to get an element of an object as a array

fn get_object<Q>(&self, k: &Q) -> Option<&Self::Object>
where Self::Key: Borrow<Q> + Hash + Eq, Q: Hash + Eq + Ord + ?Sized,

Tries to get an element of an object as a object

Object Safety

This trait is not object safe.