simd_json::prelude
Trait MutableObject

pub trait MutableObject {
    type Key: ?Sized;
    type Target;
    type Object;

    // Required methods
    fn insert<K, V>(
        &mut self,
        k: K,
        v: V,
    ) -> Result<Option<Self::Target>, AccessError>
       where Self::Key: From<K> + Hash + Eq,
             V: Into<Self::Target>;
    fn remove<Q>(&mut self, k: &Q) -> Result<Option<Self::Target>, AccessError>
       where Self::Key: Borrow<Q> + Hash + Eq,
             Q: Hash + Eq + Ord + ?Sized;
    fn get_mut<Q>(&mut self, k: &Q) -> Option<&mut Self::Target>
       where Self::Key: Borrow<Q> + Hash + Eq,
             Q: Hash + Eq + Ord + ?Sized;

    // Provided methods
    fn try_insert<K, V>(&mut self, k: K, v: V) -> Option<Self::Target>
       where Self::Key: From<K> + Hash + Eq,
             V: Into<Self::Target> { ... }
    fn try_remove<Q>(&mut self, k: &Q) -> Option<Self::Target>
       where Self::Key: Borrow<Q> + Hash + Eq,
             Q: Hash + Eq + Ord + ?Sized { ... }
}

Prelude to include needed traits Mutatability for object like values

Required Associated Types

type Key: ?Sized

The type for Object Keys

type Target

The type for Object Values

type Object

The type for Objects

Required Methods

fn insert<K, V>( &mut self, k: K, v: V, ) -> Result<Option<Self::Target>, AccessError>
where Self::Key: From<K> + Hash + Eq, V: Into<Self::Target>,

Insert into this Value as an Object. Will return an AccessError::NotAnObject if called on a Value that isn’t an object - otherwise will behave the same as HashMap::insert

Errors

Will return Err if self is not an object.

fn remove<Q>(&mut self, k: &Q) -> Result<Option<Self::Target>, AccessError>
where Self::Key: Borrow<Q> + Hash + Eq, Q: Hash + Eq + Ord + ?Sized,

Remove from this Value as an Object. Will return an AccessError::NotAnObject if called on a Value that isn’t an object - otherwise will behave the same as HashMap::remove

Errors

Will return Err if self is not an Object.

fn get_mut<Q>(&mut self, k: &Q) -> Option<&mut Self::Target>
where Self::Key: Borrow<Q> + Hash + Eq, Q: Hash + Eq + Ord + ?Sized,

Same as get but returns a mutable ref instead

Provided Methods

fn try_insert<K, V>(&mut self, k: K, v: V) -> Option<Self::Target>
where Self::Key: From<K> + Hash + Eq, V: Into<Self::Target>,

Tries to insert into this Value as an Object. If the Value isn’t an object this opoeration will return None and have no effect.

fn try_remove<Q>(&mut self, k: &Q) -> Option<Self::Target>
where Self::Key: Borrow<Q> + Hash + Eq, Q: Hash + Eq + Ord + ?Sized,

Tries to remove from this Value as an Object. If the Value isn’t an object this opoeration will return None and have no effect.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.