simd_json::prelude

Trait TypedScalarValue

pub trait TypedScalarValue {

    // Required methods
    fn is_null(&self) -> bool;
    fn is_float(&self) -> bool;
    fn is_integer(&self) -> bool;
    fn is_number(&self) -> bool;
    fn is_bool(&self) -> bool;
    fn is_i128(&self) -> bool;
    fn is_i64(&self) -> bool;
    fn is_i32(&self) -> bool;
    fn is_i16(&self) -> bool;
    fn is_i8(&self) -> bool;
    fn is_u128(&self) -> bool;
    fn is_u64(&self) -> bool;
    fn is_usize(&self) -> bool;
    fn is_u32(&self) -> bool;
    fn is_u16(&self) -> bool;
    fn is_u8(&self) -> bool;
    fn is_f64(&self) -> bool;
    fn is_f64_castable(&self) -> bool;
    fn is_f32(&self) -> bool;
    fn is_str(&self) -> bool;
    fn is_char(&self) -> bool;

}

Type checks for scalar values on a value

Required Methods

fn is_null(&self) -> bool

returns true if the current value is null

fn is_float(&self) -> bool

returns true if the current value a floatingpoint number

fn is_integer(&self) -> bool

returns true if the current value a integer number

fn is_number(&self) -> bool

returns true if the current value a number either float or intege

fn is_bool(&self) -> bool

returns true if the current value a bool

fn is_i128(&self) -> bool

returns true if the current value can be represented as a i128

fn is_i64(&self) -> bool

returns true if the current value can be represented as a i64

fn is_i32(&self) -> bool

returns true if the current value can be represented as a i32

fn is_i16(&self) -> bool

returns true if the current value can be represented as a i16

fn is_i8(&self) -> bool

returns true if the current value can be represented as a i8

fn is_u128(&self) -> bool

returns true if the current value can be represented as a u128

fn is_u64(&self) -> bool

returns true if the current value can be represented as a u64

fn is_usize(&self) -> bool

returns true if the current value can be represented as a usize

fn is_u32(&self) -> bool

returns true if the current value can be represented as a u32

fn is_u16(&self) -> bool

returns true if the current value can be represented as a u16

fn is_u8(&self) -> bool

returns true if the current value can be represented as a u8

fn is_f64(&self) -> bool

returns true if the current value can be represented as a f64

fn is_f64_castable(&self) -> bool

returns true if the current value can be cast into a f64

fn is_f32(&self) -> bool

returns true if the current value can be represented as a f64

fn is_str(&self) -> bool

returns true if the current value can be represented as a str

fn is_char(&self) -> bool

returns true if the current value can be represented as a char