Trait simd_json::prelude::ValueAsContainer

pub trait ValueAsContainer {
    type Array: Array;
    type Object: Object;

    // Required methods
    fn as_array(&self) -> Option<&Self::Array>;
    fn as_object(&self) -> Option<&Self::Object>;
}

Prelude to include needed traits Trait to allow accessing data inside a Value

Required Associated Types

type Array: Array

The array structure

type Object: Object

The object structure

Required Methods

fn as_array(&self) -> Option<&Self::Array>

Tries to represent the value as an array and returns a reference to it

fn as_object(&self) -> Option<&Self::Object>

Tries to represent the value as an object and returns a reference to it