simd_json
Module serde

serde related helper functions

Enums

SerdeConversionError

Error while converting from or to serde values

Functions

from_borrowed_value

Tries to convert a BorrowedValue into a struct that implements serde’s Deserialize interface

from_owned_value

Tries to convert a OwnedValue into a struct that implements serde’s Deserialize interface

from_reader

parses a Reader using a serde deserializer.

from_reader_with_buffers

Parses a Reader using a serde deserializer.

from_refborrowed_value

Tries to convert a &BorrowedValue into a struct that implements serde’s Deserialize interface

from_refowned_value

Tries to convert a &OwnedValue into a struct that implements serde’s Deserialize interface

from_slice

parses a byte slice using a serde deserializer. note that the slice will be rewritten in the process.

from_slice_with_buffers

Parses a byte slice using a serde deserializer. note that the slice will be rewritten in the process.

from_str⚠

Parses a str using a serde deserializer. note that the slice will be rewritten in the process and might not remain a valid utf8 string in its entirety.

from_str_with_buffers⚠

Parses a str using a serde deserializer. note that the slice will be rewritten in the process and might not remain a valid utf8 string in its entirety.

to_borrowed_value

Tries to convert a struct that implements serde’s serialize into an BorrowedValue

to_owned_value

Tries to convert a struct that implements serde’s serialize into an OwnedValue

to_string

Write a value to a string

to_string_pretty

Write a value to a string

to_vec

Write a value to a vector

to_vec_pretty

Write a value to a vector

to_writer

Write a value to a string

to_writer_pretty

Write a value to a string