simd_json

Module value

simd-json JSON-DOM value

Re-exports

pub use self::borrowed::to_value as to_borrowed_value;

pub use self::borrowed::to_value_with_buffers as to_borrowed_value_with_buffers;

pub use self::borrowed::Value as BorrowedValue;

pub use self::owned::to_value as to_owned_value;

pub use self::owned::to_value_with_buffers as to_owned_value_with_buffers;

pub use self::owned::Value as OwnedValue;

Modules

borrowed

This module holds the two dom implementations we use.

generator

Traits for serializing JSON

lazy

Lazy value, uses a tape until mutated.

owned

Owned, lifetimeless version of the value for times when lifetimes are to be avoided

prelude

Prelude for traits

tape

Tape implementation

Structs

TryTypeError

A type error thrown by the try_* functions

Enums

AccessError

An access error for ValueType

ExtendedValueType

Extended types that have no native representation in JSON

StaticNode

Static tape node

ValueType

Types of JSON values

Traits

ValueBuilder

Support of builder methods for traits.

Functions

deserialize

Parses a slice of bytes into a Value dom.

deserialize_with_buffers

Parses a slice of bytes into a Value dom.

Type Aliases

ObjectHasher

Hasher used for objects