Trait simd_json::value::generator::BaseGenerator

pub trait BaseGenerator {
    type T: Write;

    // Required methods
    fn get_writer(&mut self) -> &mut Self::T;
    fn write_min(&mut self, slice: &[u8], min: u8) -> Result<(), Error>;

    // Provided methods
    fn write(&mut self, slice: &[u8]) -> Result<(), Error> { ... }
    fn write_char(&mut self, ch: u8) -> Result<(), Error> { ... }
    fn new_line(&mut self) -> Result<(), Error> { ... }
    fn indent(&mut self) { ... }
    fn dedent(&mut self) { ... }
    fn write_string_complex(
        &mut self,
        string: &[u8],
        start: usize
    ) -> Result<(), Error> { ... }
    fn write_string(&mut self, string: &str) -> Result<(), Error> { ... }
    fn write_string_content(&mut self, string: &str) -> Result<(), Error> { ... }
    fn write_simple_string(&mut self, string: &str) -> Result<(), Error> { ... }
    fn write_simple_str_content(&mut self, string: &str) -> Result<(), Error> { ... }
    fn write_float(&mut self, num: f64) -> Result<(), Error> { ... }
    fn write_int<I>(&mut self, num: I) -> Result<(), Error>
       where I: Integer { ... }
    fn write_int128(&mut self, num: i128) -> Result<(), Error> { ... }
    fn write_uint(&mut self, num: u64) -> Result<(), Error> { ... }
    fn write_uint128(&mut self, num: u128) -> Result<(), Error> { ... }
    unsafe fn write_str_simd(&mut self, string: &mut &[u8]) -> Result<(), Error> { ... }

}

Base generator trait

Required Associated Types

type T: Write

The writer

Required Methods

fn get_writer(&mut self) -> &mut Self::T

returns teh writer

fn write_min(&mut self, slice: &[u8], min: u8) -> Result<(), Error>

write with minimum

Errors

if the write fails

Provided Methods

fn write(&mut self, slice: &[u8]) -> Result<(), Error>

Write a slice

Errors

if the write fails

fn write_char(&mut self, ch: u8) -> Result<(), Error>

Write a char

Errors

if the write fails

fn new_line(&mut self) -> Result<(), Error>

writes new line

Errors

if the write fails

fn indent(&mut self)

indents one step

fn dedent(&mut self)

dedents one step

fn write_string_complex( &mut self, string: &[u8], start: usize ) -> Result<(), Error>

Writes a string with escape sequences

Errors

if the write fails

fn write_string(&mut self, string: &str) -> Result<(), Error>

writes a string

Errors

if the write fails

fn write_string_content(&mut self, string: &str) -> Result<(), Error>

writes a string

Errors

if the write fails

fn write_simple_string(&mut self, string: &str) -> Result<(), Error>

writes a simple string (usually short and non escaped) This means we can skip the simd accelerated writing which is expensive on short strings.

Errors

if the write fails

fn write_simple_str_content(&mut self, string: &str) -> Result<(), Error>

writes a simple string content (usually short and non escaped) This means we can skip the simd accelerated writing which is expensive on short strings.

Errors

if the write fails

fn write_float(&mut self, num: f64) -> Result<(), Error>

writes a float value

Errors

if the write fails

fn write_int<I>(&mut self, num: I) -> Result<(), Error>where I: Integer,

writes an integer value

Errors

if the write fails

fn write_int128(&mut self, num: i128) -> Result<(), Error>

👎Deprecated since 0.1.5: Please use the write_int function instead

writes an integer 128 bit

Errors

if the write fails

fn write_uint(&mut self, num: u64) -> Result<(), Error>

👎Deprecated since 0.1.5: Please use the write_int function instead

writes an unsigned integer

Errors

if the write fails

fn write_uint128(&mut self, num: u128) -> Result<(), Error>

👎Deprecated since 0.1.5: Please use the write_int function instead

writes an unsigned 128bit integer

Errors

if the write fails

unsafe fn write_str_simd(&mut self, string: &mut &[u8]) -> Result<(), Error>

Writes a string with simd-acceleration

Safety

This function is unsafe because it uses simd instructions

Errors

if the write fails