simd_json::value
Module owned

Owned, lifetimeless version of the value for times when lifetimes are to be avoided

Enums

Value

Owned JSON-DOM Value, consider using the ValueTrait to access it’s content. This is slower then the BorrowedValue as a tradeoff for getting rid of lifetimes.

Functions

to_value

Parses a slice of bytes into a Value dom.

to_value_with_buffers

Parses a slice of bytes into a Value dom.

Type Aliases

Object

Representation of a JSON object