simd_json::value::prelude
Module base

Traits that provide basic interactions, they do have no auto-implementations

Traits

TypedCustomValue

Type checks for custom values on a value

TypedValue

Type information on a value

ValueAsArray

Trait to allow accessing data inside a Value

ValueAsMutArray

Mutatability for Array values

ValueAsMutObject

Mutatability for Object values

ValueAsObject

Trait to allow Value as an object

ValueAsScalar

Access to scalar value types

ValueIntoArray

A trait that specifies how to turn the Value into it’s sub types

ValueIntoObject

A trait that specifies how to turn the Value into it’s sub types

ValueIntoString

A trait that specifies how to turn the Value into it’s sub types

ValueTryAsArrayMut

try_as_array_mut access to array value types

ValueTryAsMutObject

Mutatability for Object values

Writable

A Value that can be serialized and written