Module simd_json::value::prelude::derived

Traits that have derived implementations relying on base traitsa

Traits

MutableArray

Mutatability for array like values

MutableObject

Mutatability for object like values

TypedContainerValue

Type checks for container values on a value

TypedScalarValue

Type checks for scalar values on a value

ValueArrayAccess

Access to a value as an array

ValueArrayTryAccess

Access to a value as an array with error handling

ValueObjectAccess

Access to a value as an object

ValueObjectAccessAsContainer

Access to container values in an object

ValueObjectAccessAsScalar

Access to scalar values in an object

ValueObjectAccessTryAsContainer

try_as_* access to container values in an object

ValueObjectAccessTryAsScalar

try_as_* access to scalar values in an object

ValueObjectTryAccess

try_as_* access to a value as an object

ValueTryAsContainer

try_as_* access to container value types

ValueTryAsScalar

try_as_* access to scalar value types

ValueTryIntoContainer

A trait that specifies how to turn the Value into it’s sub types with error handling

ValueTryIntoString

A trait that allows destructively turning a value into it’s string representation