simd_json::value::prelude

Trait ValueArrayAccess

pub trait ValueArrayAccess<I> {
    type Target: ?Sized;

    // Required method
    fn get_idx(&self, i: I) -> Option<&Self::Target>;
}

Access to a value as an array

Required Associated Types

type Target: ?Sized

The target for nested lookups

Required Methods

fn get_idx(&self, i: I) -> Option<&Self::Target>

Gets a ref to a value based on n index, returns None if the current Value isn’t an Array or doesn’t contain the index it was asked for.