Struct simd_json::value::tape::Value

pub struct Value<'tape, 'input>(/* private fields */)
where
    'input: 'tape;

Wrapper around the tape that allows interaction via a Value-like API.

Implementations

impl<'tape, 'input> Value<'tape, 'input>

pub fn is_array(&self) -> bool

returns true if the current value can be represented as an array

pub fn is_object(&self) -> bool

returns true if the current value can be represented as an object

impl<'tape, 'input> Value<'tape, 'input>

pub fn try_get<Q>( &self, k: &Q, ) -> Result<Option<Value<'tape, 'input>>, TryTypeError>
where str: Borrow<Q> + Hash + Eq, Q: ?Sized + Hash + Eq + Ord,

Tries to get a value based on a key, returns a TryTypeError if the current Value isn’t an Object, returns None if the key isn’t in the object

Errors

if the value is not an object

impl<'tape, 'input> Value<'tape, 'input>
where 'input: 'tape,

pub fn try_get_idx( &self, i: usize, ) -> Result<Option<Value<'tape, 'input>>, TryTypeError>

Tries to get a value based on n index, returns a type error if the current value isn’t an Array, returns None if the index is out of bounds

Errors

if the requested type doesn’t match the actual type or the value is not an object

impl<'tape, 'input> Value<'tape, 'input>
where 'input: 'tape,

pub fn as_array(&self) -> Option<Array<'tape, 'input>>

Tries to represent the value as an array and returns a reference to it

pub fn as_object(&self) -> Option<Object<'tape, 'input>>

Tries to represent the value as an array and returns a reference to it

impl<'tape, 'input> Value<'tape, 'input>

pub fn try_as_array(&self) -> Result<Array<'tape, 'input>, TryTypeError>

Tries to represent the value as an array and returns a reference to it

Errors

if the requested type doesn’t match the actual type

pub fn try_as_object(&self) -> Result<Object<'tape, 'input>, TryTypeError>

Tries to represent the value as an object and returns a reference to it

Errors

if the requested type doesn’t match the actual type

impl<'tape, 'input> Value<'tape, 'input>

pub fn get<Q>(&self, k: &Q) -> Option<Value<'tape, 'input>>
where str: Borrow<Q> + Hash + Eq, Q: ?Sized + Hash + Eq + Ord,

Gets a ref to a value based on a key, returns None if the current Value isn’t an Object or doesn’t contain the key it was asked for.

pub fn contains_key(&self, k: &str) -> bool

Checks if a Value contains a given key. This will return flase if Value isn’t an object

impl<'tape, 'input> Value<'tape, 'input>

pub fn get_idx(&self, i: usize) -> Option<Value<'tape, 'input>>

Gets a ref to a value based on n index, returns None if the current Value isn’t an Array or doesn’t contain the index it was asked for.

impl<'tape, 'input> Value<'tape, 'input>

pub fn get_array<Q>(&self, k: &Q) -> Option<Array<'tape, 'input>>
where str: Borrow<Q> + Hash + Eq, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a array

pub fn get_object<Q>(&self, k: &Q) -> Option<Object<'tape, 'input>>
where str: Borrow<Q> + Hash + Eq, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as a object

impl<'tape, 'input> Value<'tape, 'input>

pub fn try_get_array<Q>( &self, k: &Q, ) -> Result<Option<Array<'tape, 'input>>, TryTypeError>
where str: Borrow<Q> + Hash + Eq, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as an array, returns an error if it isn’t a array

Errors

if the requested type doesn’t match the actual type or the value is not an object

pub fn try_get_object<Q>( &self, k: &Q, ) -> Result<Option<Object<'tape, 'input>>, TryTypeError>
where str: Borrow<Q> + Hash + Eq, Q: ?Sized + Hash + Eq + Ord,

Tries to get an element of an object as an object, returns an error if it isn’t an object

Errors

if the requested type doesn’t match the actual type or the value is not an object