Trait simd_json::value::ValueAccess

pub trait ValueAccess {
    type Target: ValueAccess;
    type Key: Hash + Eq;
    type Array: Array
    where
        <Self::Array as Array>::Element == Self::Target;
    type Object: Object
    where
        <Self::Object as Object>::Key == Self::Key,
        <Self::Object as Object>::Element == Self::Target;

    fn value_type(&self) -> ValueType;
    fn as_bool(&self) -> Option<bool>;
    fn as_i64(&self) -> Option<i64>;
    fn as_u64(&self) -> Option<u64>;
    fn as_f64(&self) -> Option<f64>;
    fn as_str(&self) -> Option<&str>;
    fn as_array(&self) -> Option<&Self::Array>;
    fn as_object(&self) -> Option<&Self::Object>;

    fn try_as_bool(&self) -> Result<bool, TryTypeError> { ... }
    fn as_i128(&self) -> Option<i128> { ... }
    fn try_as_i128(&self) -> Result<i128, TryTypeError> { ... }
    fn try_as_i64(&self) -> Result<i64, TryTypeError> { ... }
    fn as_i32(&self) -> Option<i32> { ... }
    fn try_as_i32(&self) -> Result<i32, TryTypeError> { ... }
    fn as_i16(&self) -> Option<i16> { ... }
    fn try_as_i16(&self) -> Result<i16, TryTypeError> { ... }
    fn as_i8(&self) -> Option<i8> { ... }
    fn try_as_i8(&self) -> Result<i8, TryTypeError> { ... }
    fn as_u128(&self) -> Option<u128> { ... }
    fn try_as_u128(&self) -> Result<u128, TryTypeError> { ... }
    fn try_as_u64(&self) -> Result<u64, TryTypeError> { ... }
    fn as_usize(&self) -> Option<usize> { ... }
    fn try_as_usize(&self) -> Result<usize, TryTypeError> { ... }
    fn as_u32(&self) -> Option<u32> { ... }
    fn try_as_u32(&self) -> Result<u32, TryTypeError> { ... }
    fn as_u16(&self) -> Option<u16> { ... }
    fn try_as_u16(&self) -> Result<u16, TryTypeError> { ... }
    fn as_u8(&self) -> Option<u8> { ... }
    fn try_as_u8(&self) -> Result<u8, TryTypeError> { ... }
    fn try_as_f64(&self) -> Result<f64, TryTypeError> { ... }
    fn cast_f64(&self) -> Option<f64> { ... }
    fn try_cast_f64(&self) -> Result<f64, TryTypeError> { ... }
    fn as_f32(&self) -> Option<f32> { ... }
    fn try_as_f32(&self) -> Result<f32, TryTypeError> { ... }
    fn try_as_str(&self) -> Result<&str, TryTypeError> { ... }
    fn as_char(&self) -> Option<char> { ... }
    fn try_as_char(&self) -> Result<char, TryTypeError> { ... }
    fn try_as_array(&self) -> Result<&Self::Array, TryTypeError> { ... }
    fn try_as_object(&self) -> Result<&Self::Object, TryTypeError> { ... }
    fn get<Q>(&self, k: &Q) -> Option<&Self::Target>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get<Q>(&self, k: &Q) -> Result<Option<&Self::Target>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn contains_key<Q>(&self, k: &Q) -> bool
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_idx(&self, i: usize) -> Option<&Self::Target> { ... }
    fn try_get_idx(
        &self,
        i: usize
    ) -> Result<Option<&Self::Target>, TryTypeError> { ... }
    fn get_bool<Q>(&self, k: &Q) -> Option<bool>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_bool<Q>(&self, k: &Q) -> Result<Option<bool>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_i128<Q>(&self, k: &Q) -> Option<i128>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_i128<Q>(&self, k: &Q) -> Result<Option<i128>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_i64<Q>(&self, k: &Q) -> Option<i64>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_i64<Q>(&self, k: &Q) -> Result<Option<i64>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_i32<Q>(&self, k: &Q) -> Option<i32>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_i32<Q>(&self, k: &Q) -> Result<Option<i32>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_i16<Q>(&self, k: &Q) -> Option<i16>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_i16<Q>(&self, k: &Q) -> Result<Option<i16>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_i8<Q>(&self, k: &Q) -> Option<i8>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_i8<Q>(&self, k: &Q) -> Result<Option<i8>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_u128<Q>(&self, k: &Q) -> Option<u128>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_u128<Q>(&self, k: &Q) -> Result<Option<u128>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_u64<Q>(&self, k: &Q) -> Option<u64>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_u64<Q>(&self, k: &Q) -> Result<Option<u64>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_usize<Q>(&self, k: &Q) -> Option<usize>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_usize<Q>(&self, k: &Q) -> Result<Option<usize>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_u32<Q>(&self, k: &Q) -> Option<u32>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_u32<Q>(&self, k: &Q) -> Result<Option<u32>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_u16<Q>(&self, k: &Q) -> Option<u16>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_u16<Q>(&self, k: &Q) -> Result<Option<u16>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_u8<Q>(&self, k: &Q) -> Option<u8>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_u8<Q>(&self, k: &Q) -> Result<Option<u8>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_f64<Q>(&self, k: &Q) -> Option<f64>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_f64<Q>(&self, k: &Q) -> Result<Option<f64>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_f32<Q>(&self, k: &Q) -> Option<f32>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_f32<Q>(&self, k: &Q) -> Result<Option<f32>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_str<Q>(&self, k: &Q) -> Option<&str>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_str<Q>(&self, k: &Q) -> Result<Option<&str>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_array<Q>(
        &self,
        k: &Q
    ) -> Option<&<Self::Target as ValueAccess>::Array>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_array<Q>(
        &self,
        k: &Q
    ) -> Result<Option<&<Self::Target as ValueAccess>::Array>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn get_object<Q>(
        &self,
        k: &Q
    ) -> Option<&<Self::Target as ValueAccess>::Object>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }
    fn try_get_object<Q>(
        &self,
        k: &Q
    ) -> Result<Option<&<Self::Target as ValueAccess>::Object>, TryTypeError>
    where
        Q: Hash + Eq + Ord + ?Sized,
        Self::Key: Borrow<Q>,
        Self::Key: Hash,
        Self::Key: Eq,
    { ... }

}

Trait to allow accessing data inside a Value

Required Associated Types

type Target: ValueAccess

The target for nested lookups

type Key: Hash + Eq

The type for Objects

type Array: Array
where
<Self::Array as Array>::Element == Self::Target

The array structure

type Object: Object
where
<Self::Object as Object>::Key == Self::Key,
<Self::Object as Object>::Element == Self::Target

The object structure

Required Methods

fn value_type(&self) -> ValueType

Gets the type of the current value

fn as_bool(&self) -> Option<bool>

Tries to represent the value as a bool

fn as_i64(&self) -> Option<i64>

Tries to represent the value as an i64

fn as_u64(&self) -> Option<u64>

Tries to represent the value as an u64

fn as_f64(&self) -> Option<f64>

Tries to represent the value as a f64

fn as_str(&self) -> Option<&str>

Tries to represent the value as a &str

fn as_array(&self) -> Option<&Self::Array>

Tries to represent the value as an array and returns a refference to it

fn as_object(&self) -> Option<&Self::Object>

Tries to represent the value as an object and returns a refference to it

Provided Methods

fn try_as_bool(&self) -> Result<bool, TryTypeError>

Tries to represent the value as a bool

Errors

if the requested type doesn’t match the actual type

fn as_i128(&self) -> Option<i128>

Tries to represent the value as an i128

fn try_as_i128(&self) -> Result<i128, TryTypeError>

Tries to represent the value as a i128

Errors

if the requested type doesn’t match the actual type

fn try_as_i64(&self) -> Result<i64, TryTypeError>

Tries to represent the value as an i64

Errors

if the requested type doesn’t match the actual type

fn as_i32(&self) -> Option<i32>

Tries to represent the value as an i32

fn try_as_i32(&self) -> Result<i32, TryTypeError>

Tries to represent the value as an i32

Errors

if the requested type doesn’t match the actual type

fn as_i16(&self) -> Option<i16>

Tries to represent the value as an i16

fn try_as_i16(&self) -> Result<i16, TryTypeError>

Tries to represent the value as an i16

Errors

if the requested type doesn’t match the actual type

fn as_i8(&self) -> Option<i8>

Tries to represent the value as an i8

fn try_as_i8(&self) -> Result<i8, TryTypeError>

Tries to represent the value as an i8

Errors

if the requested type doesn’t match the actual type

fn as_u128(&self) -> Option<u128>

Tries to represent the value as an u128

fn try_as_u128(&self) -> Result<u128, TryTypeError>

Tries to represent the value as an u128

Errors

if the requested type doesn’t match the actual type

fn try_as_u64(&self) -> Result<u64, TryTypeError>

Tries to represent the value as an u64

Errors

if the requested type doesn’t match the actual type

fn as_usize(&self) -> Option<usize>

Tries to represent the value as an usize

fn try_as_usize(&self) -> Result<usize, TryTypeError>

Tries to represent the value as an usize

Errors

if the requested type doesn’t match the actual type

fn as_u32(&self) -> Option<u32>

Tries to represent the value as an u32

fn try_as_u32(&self) -> Result<u32, TryTypeError>

Tries to represent the value as an u32

Errors

if the requested type doesn’t match the actual type

fn as_u16(&self) -> Option<u16>

Tries to represent the value as an u16

fn try_as_u16(&self) -> Result<u16, TryTypeError>

Tries to represent the value as an u16

Errors

if the requested type doesn’t match the actual type

fn as_u8(&self) -> Option<u8>

Tries to represent the value as an u8

fn try_as_u8(&self) -> Result<u8, TryTypeError>

Tries to represent the value as an u8

Errors

if the requested type doesn’t match the actual type

fn try_as_f64(&self) -> Result<f64, TryTypeError>

Tries to represent the value as a f64

Errors

if the requested type doesn’t match the actual type

fn cast_f64(&self) -> Option<f64>

Casts the current value to a f64 if possible, this will turn integer values into floats.

fn try_cast_f64(&self) -> Result<f64, TryTypeError>

Tries to Casts the current value to a f64 if possible, this will turn integer values into floats and error if it isn’t possible

Errors

if the requested type doesn’t match the actual type

fn as_f32(&self) -> Option<f32>

Tries to represent the value as a f32

fn try_as_f32(&self) -> Result<f32, TryTypeError>

Tries to represent the value as a f32

Errors

if the requested type doesn’t match the actual type

fn try_as_str(&self) -> Result<&str, TryTypeError>

Tries to represent the value as a &str

Errors

if the requested type doesn’t match the actual type

fn as_char(&self) -> Option<char>

Tries to represent the value as a Char

fn try_as_char(&self) -> Result<char, TryTypeError>

Tries to represent the value as a Char

Errors

if the requested type doesn’t match the actual type

fn try_as_array(&self) -> Result<&Self::Array, TryTypeError>

Tries to represent the value as an array and returns a refference to it

Errors

if the requested type doesn’t match the actual type

fn try_as_object(&self) -> Result<&Self::Object, TryTypeError>

Tries to represent the value as an object and returns a refference to it

Errors

if the requested type doesn’t match the actual type

fn get<Q>(&self, k: &Q) -> Option<&Self::Target> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Gets a ref to a value based on a key, returns None if the current Value isn’t an Object or doesn’t contain the key it was asked for.

fn try_get<Q>(&self, k: &Q) -> Result<Option<&Self::Target>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Trys to get a value based on a key, returns a TryTypeError if the current Value isn’t an Object, returns None if the key isn’t in the object

Errors

if the value is not an object

fn contains_key<Q>(&self, k: &Q) -> bool where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Checks if a Value contains a given key. This will return flase if Value isn’t an object

fn get_idx(&self, i: usize) -> Option<&Self::Target>

Gets a ref to a value based on n index, returns None if the current Value isn’t an Array or doesn’t contain the index it was asked for.

fn try_get_idx(&self, i: usize) -> Result<Option<&Self::Target>, TryTypeError>

Tries to get a value based on n index, returns a type error if the current value isn’t an Array, returns None if the index is out of bouds

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_bool<Q>(&self, k: &Q) -> Option<bool> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a bool

fn try_get_bool<Q>(&self, k: &Q) -> Result<Option<bool>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a bool, returns an error if it isn’t bool

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_i128<Q>(&self, k: &Q) -> Option<i128> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a i128

fn try_get_i128<Q>(&self, k: &Q) -> Result<Option<i128>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a i128, returns an error if it isn’t i128

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_i64<Q>(&self, k: &Q) -> Option<i64> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a i64

fn try_get_i64<Q>(&self, k: &Q) -> Result<Option<i64>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a i64, returns an error if it isn’t a i64

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_i32<Q>(&self, k: &Q) -> Option<i32> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a i32

fn try_get_i32<Q>(&self, k: &Q) -> Result<Option<i32>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a i32, returns an error if it isn’t a i32

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_i16<Q>(&self, k: &Q) -> Option<i16> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a i16

fn try_get_i16<Q>(&self, k: &Q) -> Result<Option<i16>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a i16, returns an error if it isn’t a i16

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_i8<Q>(&self, k: &Q) -> Option<i8> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a i8

fn try_get_i8<Q>(&self, k: &Q) -> Result<Option<i8>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a i8, returns an error if it isn’t a i8

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_u128<Q>(&self, k: &Q) -> Option<u128> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a u128

fn try_get_u128<Q>(&self, k: &Q) -> Result<Option<u128>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a u128, returns an error if it isn’t a u128

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_u64<Q>(&self, k: &Q) -> Option<u64> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a u64

fn try_get_u64<Q>(&self, k: &Q) -> Result<Option<u64>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a u64, returns an error if it isn’t a u64

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_usize<Q>(&self, k: &Q) -> Option<usize> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a usize

fn try_get_usize<Q>(&self, k: &Q) -> Result<Option<usize>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a usize, returns an error if it isn’t a usize

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_u32<Q>(&self, k: &Q) -> Option<u32> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a u32

fn try_get_u32<Q>(&self, k: &Q) -> Result<Option<u32>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a u32, returns an error if it isn’t a u32

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_u16<Q>(&self, k: &Q) -> Option<u16> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a u16

fn try_get_u16<Q>(&self, k: &Q) -> Result<Option<u16>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a u16, returns an error if it isn’t a u16

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_u8<Q>(&self, k: &Q) -> Option<u8> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a u8

fn try_get_u8<Q>(&self, k: &Q) -> Result<Option<u8>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a u8, returns an error if it isn’t a u8

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_f64<Q>(&self, k: &Q) -> Option<f64> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a f64

fn try_get_f64<Q>(&self, k: &Q) -> Result<Option<f64>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a u8, returns an error if it isn’t a u8

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_f32<Q>(&self, k: &Q) -> Option<f32> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a f32

fn try_get_f32<Q>(&self, k: &Q) -> Result<Option<f32>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a f32, returns an error if it isn’t a f32

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_str<Q>(&self, k: &Q) -> Option<&str> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a str

fn try_get_str<Q>(&self, k: &Q) -> Result<Option<&str>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a str, returns an error if it isn’t a str

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_array<Q>(&self, k: &Q) -> Option<&<Self::Target as ValueAccess>::Array> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a array

fn try_get_array<Q>(
&self,
k: &Q
) -> Result<Option<&<Self::Target as ValueAccess>::Array>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as an array, returns an error if it isn’t a array

Errors

if the requested type doesn’t match the actual type or the value is not an object

fn get_object<Q>(&self, k: &Q) -> Option<&<Self::Target as ValueAccess>::Object> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as a object

fn try_get_object<Q>(
&self,
k: &Q
) -> Result<Option<&<Self::Target as ValueAccess>::Object>, TryTypeError> where
Q: Hash + Eq + Ord + ?Sized,
Self::Key: Borrow<Q>,
Self::Key: Hash,
Self::Key: Eq,

Tries to get an element of an object as an object, returns an error if it isn’t an object

Errors

if the requested type doesn’t match the actual type or the value is not an object