simd_json

Module prelude

Default trait imports;

Modules

base

Traits that provide basic interactions, they do have no auto-implementations

derived

Traits that have derived implementations relying on base traitsa

generator

Traits for serializing JSON

Structs

TryTypeError

A type error thrown by the try_* functions

Enums

AccessError

An access error for ValueType

ExtendedValueType

Extended types that have no native representation in JSON

StaticNode

Static tape node

ValueType

Types of JSON values

Traits

ArrayMut

Mutability functions for a value array

ArrayTrait

A trait for the minimal common functionality of a vale array

BaseGenerator

Base generator trait

Indexed

Trait for indexing into an array

IndexedMut

Trait for indexing into an array

MutableArray

Mutatability for array like values

MutableObject

Mutatability for object like values

MutableValueArrayAccess

Mutatability for array like values

ObjectMut

A mutable value Object

ObjectTrait

A trait for the minimal common functionality of a vale object

TypedArrayValue

Type checks for array values on a value

TypedCustomValue

Type checks for custom values on a value

TypedObjectValue

Type checks for object values on a value

TypedScalarValue

Type checks for scalar values on a value

TypedValue

Type information on a value

ValueArrayAccess

Access to a value as an array

ValueArrayTryAccess

Access to a value as an array with error handling

ValueAsArray

Trait to allow accessing data inside a Value

ValueAsMutArray

Mutatability for Array values

ValueAsMutObject

Mutatability for Object values

ValueAsObject

Trait to allow Value as an object

ValueAsScalar

Access to scalar value types

ValueBuilder

Support of builder methods for traits.

ValueIntoArray

A trait that specifies how to turn the Value into it’s sub types

ValueIntoObject

A trait that specifies how to turn the Value into it’s sub types

ValueIntoString

A trait that specifies how to turn the Value into it’s sub types

ValueObjectAccess

Access to a value as an object

ValueObjectAccessAsArray

Access to array values in an object

ValueObjectAccessAsObject

Access to object values in an object

ValueObjectAccessAsScalar

Access to scalar values in an object

ValueObjectAccessTryAsArray

try_get_array access to object values in an object

ValueObjectAccessTryAsObject

try_get_object access to object values in an object

ValueObjectAccessTryAsScalar

try_as_* access to scalar values in an object

ValueObjectTryAccess

try_as_* access to a value as an object

ValueTryAsArray

try_as_* access to array value types

ValueTryAsArrayMut

try_as_array_mut access to array value types

ValueTryAsMutObject

Mutatability for Object values

ValueTryAsObject

try_as_* access to object value types

ValueTryAsScalar

try_as_* access to scalar value types

ValueTryIntoArray

A trait that specifies how to turn the Value into it’s sub types with error handling

ValueTryIntoObject

A trait that specifies how to turn the Value into it’s sub types with error handling

ValueTryIntoString

A trait that allows destructively turning a value into it’s string representation

Writable

A Value that can be serialized and written