Module simd_json::prelude

Default trait imports;

Modules

base

Prelude to include needed traits Traits that provide basic interactions, they do have no auto-implementations

derived

Prelude to include needed traits Traits that have derived implementations relying on base traitsa

generator

Prelude to include needed traits Traits for serializing JSON

Structs

TryTypeError

Prelude to include needed traits A type error thrown by the try_* functions

Enums

AccessError

Prelude to include needed traits An access error for ValueType

ExtendedValueType

Prelude to include needed traits Extended types that have no native representation in JSON

StaticNode

Prelude to include needed traits Static tape node

ValueType

Prelude to include needed traits Types of JSON values

Traits

ArrayMut

Prelude to include needed traits Mutability functions for a value array

ArrayTrait

Prelude to include needed traits A trait for the minimal common functionality of a vale array

BaseGenerator

Prelude to include needed traits Base generator trait

MutableArray

Prelude to include needed traits Mutatability for array like values

MutableObject

Prelude to include needed traits Mutatability for object like values

ObjectMut

Prelude to include needed traits A mutable value Object

ObjectTrait

Prelude to include needed traits A trait for the minimal common functionality of a vale object

TypedContainerValue

Prelude to include needed traits Type checks for container values on a value

TypedCustomValue

Prelude to include needed traits Type checks for custom values on a value

TypedScalarValue

Prelude to include needed traits Type checks for scalar values on a value

TypedValue

Prelude to include needed traits Type information on a value

ValueArrayAccess

Prelude to include needed traits Access to a value as an array

ValueArrayTryAccess

Prelude to include needed traits Access to a value as an array with error handling

ValueAsContainer

Prelude to include needed traits Trait to allow accessing data inside a Value

ValueAsMutContainer

Prelude to include needed traits Mutatability for container values

ValueAsScalar

Prelude to include needed traits Access to scalar value types

ValueBuilder

Prelude to include needed traits Support of builder methods for traits.

ValueIntoContainer

Prelude to include needed traits A trait that specifies how to turn the Value into it’s sub types

ValueIntoString

Prelude to include needed traits A trait that specifies how to turn the Value into it’s sub types

ValueObjectAccess

Prelude to include needed traits Access to a value as an object

ValueObjectAccessAsContainer

Prelude to include needed traits Access to container values in an object

ValueObjectAccessAsScalar

Prelude to include needed traits Access to scalar values in an object

ValueObjectAccessTryAsContainer

Prelude to include needed traits try_as_* access to container values in an object

ValueObjectAccessTryAsScalar

Prelude to include needed traits try_as_* access to scalar values in an object

ValueObjectTryAccess

Prelude to include needed traits try_as_* access to a value as an object

ValueTryAsContainer

Prelude to include needed traits try_as_* access to container value types

ValueTryAsScalar

Prelude to include needed traits try_as_* access to scalar value types

ValueTryIntoContainer

Prelude to include needed traits A trait that specifies how to turn the Value into it’s sub types with error handling

ValueTryIntoString

Prelude to include needed traits A trait that allows destructively turning a value into it’s string representation

Writable

Prelude to include needed traits A Value that can be serialized and written