Module simd_json::prelude

Default trait imports;

Enums

AccessError

Prelude to include needed traits An access error for ValueType

ValueType

Prelude to include needed traits Types of JSON values

Traits

Array

Prelude to include needed traits Functions guaranteed for any array object

Builder

Prelude to include needed traits Support of builder methods for traits.

Mutable

Prelude to include needed traits Mutatability for values

Object

Prelude to include needed traits A JSON Object

ValueAccess

Prelude to include needed traits Trait to allow accessing data inside a Value

ValueTrait

Prelude to include needed traits The Value exposes common interface for values, this allows using both BorrowedValue and OwnedValue nearly interchangable

Writable

Prelude to include needed traits A Value that can be serialized and written