Module simd_json::value

simd-json JSON-DOM value

Re-exports

pub use self::borrowed::to_value as to_borrowed_value;

pub use self::borrowed::to_value_with_buffers as to_borrowed_value_with_buffers;

pub use self::borrowed::Value as BorrowedValue;

pub use self::owned::to_value as to_owned_value;

pub use self::owned::to_value_with_buffers as to_owned_value_with_buffers;

pub use self::owned::Value as OwnedValue;

Modules

borrowed

This module holds the two dom implementations we use. We distinguish between owned and borrowed. The difference being is that the borrowed value will use &str as its string type, referencing the input, while owned will allocate a new String for each value.

generator

Traits for serializing JSON

owned

Owned, lifetimeless version of the value for times when lifetimes are to be avoided

prelude

Prelude for traits

tape

Tape implementation

Structs

TryTypeError

A type error thrown by the try_* functions

Enums

AccessError

An access error for ValueType

ExtendedValueType

Extended types that have no native representation in JSON

StaticNode

Static tape node

ValueType

Types of JSON values

Traits

ValueBuilder

Support of builder methods for traits.

Functions

deserialize

Parses a slice of bytes into a Value dom. This function will rewrite the slice to de-escape strings. As we reference parts of the input slice the resulting dom has the same lifetime as the slice it was created from.

deserialize_with_buffers

Parses a slice of bytes into a Value dom. This function will rewrite the slice to de-escape strings. As we reference parts of the input slice the resulting dom has the same lifetime as the slice it was created from.

Type Aliases

ObjectHasher

Hasher used for objects