Module simd_json::value::owned

Owned, lifetimeless version of the value for times when lifetimes are to be avoided

Enums

Value

Owned JSON-DOM Value, consider using the ValueTrait to access it’s content. This is slower then the BorrowedValue as a tradeoff for getting rid of lifetimes.

Functions

to_value

Parses a slice of bytes into a Value dom. This function will rewrite the slice to de-escape strings. We do not keep any references to the raw data but re-allocate owned memory wherever required thus returning a value without a lifetime.

to_value_with_buffers

Parses a slice of bytes into a Value dom. This function will rewrite the slice to de-escape strings. We do not keep any references to the raw data but re-allocate owned memory wherever required thus returning a value without a lifetime.

Type Definitions

Object

Representation of a JSON object