Module simd_json::value::prelude

Prelude for traits

Modules

base

Traits that provide basic interactions, they do have no auto-implementations

derived

Traits that have derived implementations relying on base traitsa

Structs

TryTypeError

A type error thrown by the try_* functions

Enums

AccessError

An access error for ValueType

ExtendedValueType

Extended types that have no native representation in JSON

StaticNode

Static tape node

ValueType

Types of JSON values

Traits

ArrayMut

Mutability functions for a value array

ArrayTrait

A trait for the minimal common functionality of a vale array

BaseGenerator

Base generator trait

MutableArray

Mutatability for array like values

MutableObject

Mutatability for object like values

ObjectMut

A mutable value Object

ObjectTrait

A trait for the minimal common functionality of a vale object

TypedContainerValue

Type checks for container values on a value

TypedCustomValue

Type checks for custom values on a value

TypedScalarValue

Type checks for scalar values on a value

TypedValue

Type information on a value

ValueArrayAccess

Access to a value as an array

ValueArrayTryAccess

Access to a value as an array with error handling

ValueAsContainer

Trait to allow accessing data inside a Value

ValueAsMutContainer

Mutatability for container values

ValueAsScalar

Access to scalar value types

ValueBuilder

Support of builder methods for traits.

ValueIntoContainer

A trait that specifies how to turn the Value into it’s sub types

ValueIntoString

A trait that specifies how to turn the Value into it’s sub types

ValueObjectAccess

Access to a value as an object

ValueObjectAccessAsContainer

Access to container values in an object

ValueObjectAccessAsScalar

Access to scalar values in an object

ValueObjectAccessTryAsContainer

try_as_* access to container values in an object

ValueObjectAccessTryAsScalar

try_as_* access to scalar values in an object

ValueObjectTryAccess

try_as_* access to a value as an object

ValueTryAsContainer

try_as_* access to container value types

ValueTryAsScalar

try_as_* access to scalar value types

ValueTryIntoContainer

A trait that specifies how to turn the Value into it’s sub types with error handling

ValueTryIntoString

A trait that allows destructively turning a value into it’s string representation

Writable

A Value that can be serialized and written