Module simd_json::value::prelude

Prelude for traits

Structs

TryTypeError

A type error thrown by the try_* functions

Enums

AccessError

An access error for ValueType

ExtendedValueType

Extended types that have no native representation in JSON

ValueType

Types of JSON values

Traits

Array

Functions guaranteed for any array object

BaseGenerator

Base generator trait

Builder

Support of builder methods for traits.

Mutable

Mutatability for values

Object

A JSON Object

Value

The Value exposes common interface for values, this allows using both BorrowedValue and OwnedValue nearly interchangable

ValueAccess

Trait to allow accessing data inside a Value

Writable

A Value that can be serialized and written